ContextoCOMPLETO

+++ AcquaDiCane/Models/ReciboDePago.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public DateTime FechaEmision { get; set; }
-        public decimal Monto { get; set; }
+        public double Monto { get; set; }
         public int PagoId { get; set; }
         public Pago Pago { get; set; }
     }
